pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tender-intelligence-agent"
version = "0.1.0"
description = "MCP server for AI-powered tender qualification"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "mcp>=1.2.0",
  "openai>=1.52.0",
  "pydantic>=2.8.2",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.2",
  "requests>=2.32.3"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "ruff>=0.6.8"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
pythonpath = ["src"]
